Adds App.js as a common base for ios and android
The App is simply exported to ios and android,
this ensures we won't have to copy and paste
code into both files, and it will be easier to
identify the components which are unique to
android and ios. Also, other minor changes were
made.
