+++ src/css/styles/style.js
@@ -13,4 +13,4 @@ const styles = StyleSheet.create({
   },
 });
 
-module.exports = styles;
+export default styles;
